Added test using a WordTokenizer.
